// src/shared/api/generated/models/projectMembershipListResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectMembershipResponseDto } from "./projectMembershipResponseDto";

export interface ProjectMembershipListResponseDto {
  items: ProjectMembershipResponseDto[];
  /** @nullable */
  nextCursor: string | null;
}